Do not try to mount using /etc/fstab.kdump unless it exists
When using dracut, this file is in fact merged with /etc/fstab and
handled by dracut itself.

+++ init/save_dump.sh
@@ -111,6 +111,8 @@ function mount_all()
 {
     local ret=0
 
+    test -f /etc/fstab.kdump || return 0
+
     if [ -f /etc/fstab ] ; then
         mv /etc/fstab /etc/fstab.orig
     fi
